refactor components to sealed classes and add navigation and history features

+++ app/src/main/java/io/droidevs/calculatorplus/domain/expression/NumberExpression.kt
@@ -3,6 +3,6 @@ package io.droidevs.calculatorplus.domain.expression
 import io.droidevs.calculatorplus.domain.expression.Expression
 import java.math.BigDecimal
 
-class NumberExpression(private val value: Double) : Expression() {
-    override fun evaluate(): BigDecimal = BigDecimal(value)
+class NumberExpression(private val value: BigDecimal) : Expression() {
+    override fun evaluate(): BigDecimal = value
 }
